+++ crawler-commons/src/main/java/achecrawler/crawler/crawlercommons/fetcher/FetchedResult.java
@@ -1,11 +1,10 @@
 /**
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *  
+ * Copyright 2016 Crawler-Commons
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
  *     http://www.apache.org/licenses/LICENSE-2.0
  * 
  * Unless required by applicable law or agreed to in writing, software
@@ -40,7 +39,7 @@ public class FetchedResult {
     private Payload _payload;
 
     public FetchedResult(String baseUrl, String redirectedUrl, long fetchTime, Headers headers, byte[] content, String contentType, int responseRate, Payload payload, String newBaseUrl,
-        int numRedirects, String hostAddress, int statusCode, String reasonPhrase) {
+                    int numRedirects, String hostAddress, int statusCode, String reasonPhrase) {
         _payload = payload;
 
         if (baseUrl == null) {
@@ -146,7 +145,7 @@ public class FetchedResult {
      * . The order of the report is based on the logical population of
      * FetchedReport entities as per a non-public algorithm within
      * {@link achecrawler.crawler.crawlercommons.fetcher.http.SimpleHttpFetcher}.
-     *
+     * 
      * @return returns a String report of the FetchedResult.
      */
     public String report() {